Add \Device\KsecDD emulation and mask 32-bit syscall arguments
- Emulate \Device\KsecDD (RNG + memory encrypt/decrypt) so the bcrypt/CNG
  DllMain succeeds during Windows PE process initialization.
- Mask ULONG syscall arguments that were read as 64-bit; some ntdll builds
  leave garbage in the upper 32 bits, which broke handler argument checks.

+++ Brovan/Core/Emulation/OS/Windows/Files/NtCreateFile.cs
@@ -31,7 +31,7 @@ namespace Brovan.Core.Emulation.OS.Windows
         private NTSTATUS Handle64(BinaryEmulator Instance)
         {
             ulong FileHandlePtr = Instance.WinHelper.GetArg64(0);
-            ulong DesiredAccess = Instance.WinHelper.GetArg64(1);
+            ulong DesiredAccess = (uint)Instance.WinHelper.GetArg64(1);
             ulong ObjectAttributesPtr = Instance.WinHelper.GetArg64(2);
             ulong IoStatusBlockPtr = Instance.WinHelper.GetArg64(3);
             uint CreateDisposition = (uint)Instance.WinHelper.GetArg64(7);
